Handle windows path
# Windows handling

1. source does not work on windows, and on windows you need to directly execute the command without the source prefix
2. On windows the path uses backslash instead of slash

# Added a launch script

Not enough to be able to just install, because the user needs to be able to launch it again later on. Of course, you could let the user run the whole install script whenever they launch but that's overkill, so separated out a "launch.json" script that only takes care of the launching part.

1. By separating out the launch script from install script, you can automatically start langflow when Pinokio starts. Updated pinokio.js to set the "start" script. This will execute the "launch.json" when Pinokio starts
2. To make this possible, we need the launch.json script to access the activate_command set from the install script (start.json). So instead of using the local variable, use the "self" variable which persists data to JSON files in the same repository. In this case we are persisting the attributes to "config.json" from "start.json", and then later accessing them using "self.config" variable from everywhere (these are automatically imported in all template expressions regardless of which file)

python: use python instead of python3 (on the latest standalone python3 binary on windows, the python binary is named python instead of python3, assuming everything is python)