After starting the server need to go into standby mode
Everything started by pinokio script gets shut down and garbage collected once the script finishes executing.

This means if you started a server using the script, it will be shut down when the script finishes running.

To avoid this, all you need to do is make sure the script never finishes. You can do this with "process.wait", which effectively keeps the script on a standby mode.